Reopen #32 that was closed due to developer Github account refactoring.
Small and minimal extensions to DumpXYZ. Now can write standard and
non standard XYZ files (e.g. as fastwritexyz). 
Printing of particle type is not hard coded but 
retrieved and one can give the option to store pid
and/or velocities (default set to standard XYZ format).
Precision for printing doubles as fastwritexyx but 
can be changed. Changed fastwritexyz
and others in DumpConfigurations.py accordingly.