Reverted some of the changes in #98 in ConfigurationsExt::gather()
Reworked the retrieval of the particle type as done
in DumpGRO. The problem was that DumpGRO, etc. rely
on config.nextProperties() and with my changes it would 
have printed a wrong line and required an ugly to see but trivial
fix in ConfigurationExtIterator::nextProperties().
In this way it still extend DumpXYZ, get the right particle type
as DumpGRO and doesn't break any other class.