Passing const std::vector by reference instead of array.
This way we can perform the check if the size of the passed
vector matches the dimensions that we want to set for RealND
RealND(dim, vector). Otherwise we could have RealND(dim, vector[])
but then we loose info on the vector[] size unless we pass it too.
Before it was broken as its signature was
RealND(dim, vector[3]). Problems if dim > 3. Not sure this constructor
is used anywhere in the codebase.